feat(orga): moved usecases and their tests
